# openapi-doc/src/main/spec/one-stop-openapi.yaml
openapi: "3.0.2"
info:
  version: "1.0.0"
  title: "One Stop API"
  description: "One Stop API"
servers:
  - url: http://hillayes.com/api/v1
    description: Main (production) server
  - url: http://localhost/api/v1
    description: Development server
...
paths:
  /users:
    get:
      tags: [ "users" ]
      summary: "Get all users"
      description: "Get a paginated list of all users."
      operationId: "getAllUsers"
      parameters:
        - name: "page"
          in: "query"
          description: "Page number - zero indexed"
          required: false
          schema:
            type: integer
            format: int32
        - name: "page-size"
          in: "query"
          description: "Number of items per page"
          required: false
          schema:
            type: integer
            format: int32
            default: 20
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
        "401":
          description: "Unauthorized"
        "500":
          description: "Internal server error"
  /users/{id}:
    get:
      tags: [ "users" ]
      summary: "Get user by id"
      description: "Returns the details of the identified user"
      operationId: "getUserById"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of user to return"
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "Unauthorized"
        "404":
          description: "User not found"
        "500":
          description: "Internal server error"

    put:
      tags: [ "users" ]
      summary: "Update user"
      description: |
        Updates the user with the given id. Used by adminstrator to update
        a user's details.
      operationId: "updateUser"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of user to update"
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: "Updated user object"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: "Unauthorized"
        "404":
          description: "User not found"
        "500":
          description: "Internal server error"

    delete:
      tags: [ "users" ]
      summary: "Delete user"
      description: "Deletes the identified user. Admin only."
      operationId: "deleteUser"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of user to delete"
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: "successful operation"
        "401":
          description: "Unauthorized"
        "404":
          description: "User not found"
        "500":
          description: "Internal server error"
  /users/onboard/register:
    post:
      tags: [ "users" ]
      summary: "Register new user"
      description: "Registers a user's email, starting the onboarding process."
      operationId: "registerUser"
      requestBody:
        description: "Updated profile object"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterRequest"
      responses:
        "204":
          description: "successful operation"
        "409":
          description: "Email already in use."
        "500":
          description: "Internal server error"
  /users/onboard/acknowledge:
    get:
      tags: [ "users" ]
      summary: "Acknowledge user receipt of registration email"
      description: |
        Confirms the user has received the registration email, verifying their email
        address. The user will be redirected to the user confirmation page to allow a
        username and password to be selected.
      operationId: "acknowledgeUser"
      parameters:
        - name: "magicToken"
          in: "path"
          description: |
            The magicToken emailed to the user on registration. This identifies the
            user and verifies the email receipt. It expires after a configured duration,
            and is revoked once used.
          required: true
          schema:
            type: string
            example: bdfovw78y1fhjda87sdqus
      responses:
        "307":
          description: "redirected to user profile page"
        "401":
          description: "Token not valid or has expired"
        "500":
          description: "Internal server error"
  /users/onboard/complete:
    post:
      tags: [ "users" ]
      summary: "Completes the user's onboarding"
      description: |
        Sets a user's username and password, completing the onboarding process.
      operationId: "onboardUser"
      requestBody:
        description: "Updated profile object"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCompleteRequest"
      responses:
        "201":
          description: "Onboarding complete"
        "401":
          description: "Token not valid or has expired"
        "500":
          description: "Internal server error"
...
components:
  schemas:
    UserRegisterRequest:
      type: object
      required:
        - email
      properties:
        email:
          description: "The email address the new user wishes to register with"
          type: string
          format: email
    UserProfileRequest:
      type: object
      required:
        - username
        - givenName
        - email
      properties:
        username:
          type: string
          description: The user's selected username. Will be verified to ensure it is unique.
          example: jdoe
        title:
          type: string
          description: The user's title
          example: Mr
        givenName:
          type: string
          description: The user's given name
          example: John
        familyName:
          type: string
          description: The user's family name
          example: Doe
        preferredName:
          type: string
          description: The user's preferred name
          example: John
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 255
          example: john.doe@work.com
          description: The creditor user's email address
        phone:
          type: string
          description: The user's phone number
          example: 0123456789
    UserRole:
      type: string
      description: A authorisation role assigned to user
      enum:
        - user
        - admin
    UserUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/UserProfileRequest"
        - type: object
          required:
            - username
            - givenName
            - email
            - roles
          properties:
            roles:
              type: array
              items:
                $ref: "#/components/schemas/UserRole"
    UserCompleteRequest:
      allOf:
        - $ref: "#/components/schemas/UserProfileRequest"
        - type: object
          required:
            - username
            - password
            - givenName
            - email
          properties:
            password:
              type: string
              description: |
                The user's new password. Must:
                  - be at least 12 characters
                  - include at least one upper-case character
                  - include at least one lower-case character
                  - include at least one alpha character
                  - include at least one numeric character
                  - include at least one non-alpha-numeric character
              example: Simple-Banking-216
...
    UserProfileResponse:
      type: object
      required:
        - username
        - givenName
        - email
      properties:
        username:
          type: string
          description: The user's username
          example: jdoe
        title:
          type: string
          description: The user's title
          example: Mr
        givenName:
          type: string
          description: The user's given name
          example: John
        familyName:
          type: string
          description: The user's family name
          example: Doe
        preferredName:
          type: string
          description: The user's preferred name
          example: John
        email:
          type: string
          format: email
          minLength: 3
          maxLength: 255
          example: john.doe@work.com
          description: The creditor user's email address
        phone:
          type: string
          description: The user's phone number
          example: 0123456789
        dateCreated:
          type: string
          format: date-time
          description: The date the user was created
          example: 2019-01-01T00:00:00.000Z
        dateOnboarded:
          type: string
          format: date-time
          description: The date the user was onboarded
          example: 2019-01-01T00:00:00.000Z
        dateBlocked:
          type: string
          format: date-time
          description: The date the user account was blocked
          example: 2019-01-01T00:00:00.000Z
    UserResponse:
      allOf:
        - $ref: "#/components/schemas/UserProfileResponse"
        - type: object
          required:
            - id
            - username
            - givenName
            - email
            - roles
          properties:
            id:
              type: string
              format: uuid
              description: The Creditor Id
              example: 07061d76-74f3-4163-aa07-fb6909681f81
            roles:
              type: array
              items:
                type: string
                description: The authorisation roles that the user holds
                example: admin
...
    PageLinks:
      type: object
      required:
        - first
        - last
      properties:
        first:
          type: string
          format: uri
          description: The URL of the first page of items
          example: https://api.example.com/accounts?page=0&page-size=19
        next:
          type: string
          format: uri
          description: The URL of the next page of items
          example: https://api.example.com/accounts?page=3&page-size=19
        previous:
          type: string
          format: uri
          description: The URL of the previous page of items
          example: https://api.example.com/accounts?page=1&page-size=19
        last:
          type: string
          format: uri
          description: The URL of the last page of items
          example: https://api.example.com/accounts?page=15&page-size=19
    PaginatedItems:
      type: object
      required:
        - total
        - count
        - page
        - pageSize
        - links
      properties:
        total:
          type: integer
          format: int64
          description: The total number of items in the list
          example: 292
        count:
          type: integer
          format: int32
          description: The number of items in the page
          example: 20
        page:
          type: integer
          format: int32
          description: The current page number
          example: 2
        pageSize:
          type: integer
          format: int32
          description: The requested number of items per page
          example: 25
        links:
          $ref: "#/components/schemas/PageLinks"
...
    PaginatedUsers:
      allOf:
        - $ref: "#/components/schemas/PaginatedItems"
        - type: object
          properties:
            items:
              type: array
              description: The list of users
              items:
                $ref: "#/components/schemas/UserResponse"
...